backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
transformers==4.41.2
torch==2.3.1
newspaper3k==0.2.8
trafilatura==1.9.0
keybert==0.8.5
gunicorn==22.0.0
requests==2.32.3
lxml_html_clean
